requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
openai
python-dotenv
scikit-learn
matplotlib
seaborn
pandas
tqdm